move code to zerohint_others
